Carrito 99%
Falta agregar casos de cuando sea kg o lts o asi pero lo demas ya esta, intentare hacer ventas

+++ Punto de Venta/SqlScripts/Tablas.sql
@@ -465,7 +465,7 @@ create table VentaTemporal
   CodigoProducto int not null,
   NombreProducto varchar(50) not null,
   ExistenciaProducto varchar(200) not null,
-  PrecioUnitario int not null,
+  PrecioUnitario money not null,
   FechaVenta date null,
   Caja int null,
   CantidadAllevar int not null
